requirements.txt
channels
appdirs==1.4.3
attrs==19.1.0
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.2
django-background-tasks==1.2.0
django-compat==1.0.15
djangorestframework==3.9.2
geographiclib==1.49
geopy==1.19.0
idna==2.8
isodate==0.6.0
kavenegar==1.1.2
lxml==4.3.3
paho-mqtt==1.4.0
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
zeep==3.3.1
